fix: make the `records` type inference in `useTable` call work
Fixes #599

I'm not sure the exact cause of the original issue.

But my guess is that the `computed()` call complicated the type
inference process significantly - especially for the `cell` field, which
are return values of callback functions that are themselves return
values of a generic function. TypeScript loses track of the connection
between the several `R` generic types in the `useTable` call and
therefore throwing an error.

By only using `computed()` in the lowert necessary level, we can
simplify the type inference process and make it work.