refactor(input): re-export Size & Color types from SInputBase
closes #623

I also updated the `CheckColor` type to be a re-export of `Color` as the
values always match.

`SInputImage`, `SInputFileUpload`, and other non-input components have
different sizes than `SInputBase` so they are kept as-is.

There's a `Color` type export in `SInputSegments` that has an additional
`default` value. But I don't see this type used anywhere after searching
the codebase, so I guess it's redundant and removed the type completely.